requirements.txt
fastapi
uvicorn
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
pymupdf
python-docx
openai
langchain